+++ R/zzz.R
@@ -17,15 +17,20 @@ networkx <- NULL
 .onAttach <- function(libname, pkgname) {
   # Check for kuzu and pandas and provide a helpful message if they're not found
   if (interactive()) {
-    pkgs <- c("kuzu", "pandas")
+    pkgs <- c("kuzu", "pandas", "networkx")
     installed_status <- sapply(pkgs, reticulate::py_module_available)
     missing_pkgs <- names(installed_status[!installed_status])
 
     if (length(missing_pkgs) > 0) {
+      install_cmd <- sprintf(
+        'reticulate::py_install(c("%s"), pip = TRUE)',
+        paste(missing_pkgs, collapse = '", "')
+      )
       msg <- paste(
         "Required Python packages are not installed:",
         paste(missing_pkgs, collapse = ", "),
-        "\nPlease run `kuzuR::install_kuzu()` to install them."
+        "\nPlease run the following command to install them:\n",
+        install_cmd
       )
       packageStartupMessage(msg)
     }